refactor #step2 - 비즈니스 로직에 해당되는 코드를 domain 패키지로 옮깁니다
